Update file refs and kotlin organization
